Add `Tokenizable` module
`Knock::Tokenizable` can be included in the entity model (e.g. User) to
provide token serialization and deserialization.

Works out of the box with the same sensible default payload format and
fetching method using `#find`.

Or the methods can be overriden as usual.

This brings some convenience to implement a custom token creation
endpoint without relying on Knock's AuthTokenController which I'm trying
to move away from.

This allows for the possibility to make Knock even simpler by packaging
it as a simple gem instead of Rails engine. Less magic, more simplicity
and ease of customization. Just as easy to get started with clear
documentation.