v2.2 (#248)
* Change Rails dependency to 5+

* Remove Rails 4 gemfile

* Remove generators

The generator is used to created very simple files. They're adding
complexity to the gem without providing much value.

More details should be added to the README instead.

* Fix deprecations warnings

* Update README

Give more details about initializer in the Readme to compensate for the
removal of generator.

* Ensure version respect Zeitwerk loading convention

* Update ruby version in travis

* Update Changelog

* Add rubocop configuration

* Minor refactoring of authenticable module

- Implement `respond_to_missing?`
- Fix some minor styling and convention issues

* Update CHANGELOG

* Add `Tokenizable` module

`Knock::Tokenizable` can be included in the entity model (e.g. User) to
provide token serialization and deserialization.

Works out of the box with the same sensible default payload format and
fetching method using `#find`.

Or the methods can be overriden as usual.

This brings some convenience to implement a custom token creation
endpoint without relying on Knock's AuthTokenController which I'm trying
to move away from.

This allows for the possibility to make Knock even simpler by packaging
it as a simple gem instead of Rails engine. Less magic, more simplicity
and ease of customization. Just as easy to get started with clear
documentation.

* Feature/multiple token lifetime (#251)

* fixing default code style

* allowing gem's user to change token lifetime

* changing the default to user

* wip applying changes

* wip apply suggestions from code review

* wip apply suggestions from code review (README.md)

* apply suggestions from code review (README.md)

* test added and entity class access changed

* pry added to the project

* tests for default user and superuser

* removing pry and fixing documentation

* rewriting test message

Co-authored-by: Renata Marques <renata.nrmarques@gmail.com>